cleaned up server, pipelines, and strategies
