+++ test/test.js
@@ -8,6 +8,7 @@ var q = require('q'),
     module.exports.test = function() {
         test_bad_pipelines();
         test_pipelines();
+        test_cloud_strategies();
     };
 
     var PDF_BIN = 'XXXpdfcontentXXX',
@@ -34,6 +35,37 @@ var q = require('q'),
         });
     }
 
+    function test_cloud_strategies() {
+
+        cloud_strategies.AS_ARRAY.map(test_strategy);
+        function test_strategy(strategy) {
+            var cloud_hellos = strategy.make_cloud_store('hellos'),
+                cloud_create_hello = cloud_hellos.make_create_function('create-hello', cloud_hellos),
+                post_to_cloud_create_hello = cloud_create_hello.trigger_by_post(),
+                cloud_helloworld = strategy.make_cloud_function('helloworld', helloworld),
+                post_to_helloworld = cloud_helloworld.trigger_by_post()
+                ;
+            post_to_helloworld({wat: 'I am that I am'})
+                .then(function(message) {
+                    console.log('PASS post_to_helloworld:', message);
+                }, function(error) {
+
+                    console.log('FAIL post_to_helloworld:', error);
+                })
+            ;
+            function helloworld(data, context) {
+                post_to_cloud_create_hello(data)
+                    .then(function(message) {
+                        console.log('PASS post_to_cloud_create_hello:', message);
+                        context.success(message);
+                    }, function(error) {
+                        console.log('FAIL post_to_cloud_create_hello:', error);
+                        context.failure(error);
+                    })
+                ;
+            }
+        }
+    }
     function expect_success(label, fn) {
         var out;
         try {
@@ -65,3 +97,5 @@ var q = require('q'),
 })();
 
 module.exports.test();
+
+
